+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- sharing controls on item detail view in dummy
+- added `surge` and `github` environment settings
+- updated `npm run deploy` to use the `github` environment
+
 ## [0.8.5]
 ### Added
 - `item-service.uploadResource(itemId, owner, file, replace = false)` optional param so we can replace resources